requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
charset-normalizer==3.4.0
contourpy==1.3.0
cycler==0.12.1
distro==1.9.0
et_xmlfile==2.0.0
fake-useragent==1.5.1
filelock==3.16.1
fonttools==4.54.1
fsspec==2024.10.0
futures==3.0.5
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
Jinja2==3.1.4
jiter==0.7.0
joblib==1.4.2
kiwisolver==1.4.7
llvmlite==0.43.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mpmath==1.3.0
networkx==3.4.2
numba==0.60.0
numpy==2.0.2
openai==1.54.3
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pillow==11.0.0
pydantic==2.9.2
pydantic_core==2.23.4
PyMuPDF==1.24.13
pynndescent==0.5.13
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
regex==2024.11.6
requests==2.32.3
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
six==1.16.0
sniffio==1.3.1
sympy==1.13.1
threadpoolctl==3.5.0
tiktoken==0.8.0
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tqdm==4.67.0
typing_extensions==4.12.2
tzdata==2024.2
umap-learn==0.5.7
urllib3==2.2.3
xgboost==2.1.2
